# packages/daemon/openapi.yaml
openapi: 3.0.3
info:
  title: orch8 API
  description: |
    REST + WebSocket API for the orch8 orchestration platform.

    ## Authentication

    Two authentication paths are supported on all endpoints (except `GET /api/health`):

    **Agent Authentication** — Provide both `X-Agent-Id` and `X-Project-Id` headers. The agent is validated against the database.

    **Admin Authentication** — Requests from localhost (`127.0.0.1`, `::1`) without agent headers are treated as admin. Admins bypass all permission checks.
  version: 0.0.0
  license:
    name: Private
servers:
  - url: http://localhost:3847
    description: Local daemon
...
components:
  securitySchemes:
    AgentAuth:
      type: apiKey
      in: header
      name: X-Agent-Id
      description: Agent ID header (must be paired with X-Project-Id)
    ProjectScope:
      type: apiKey
      in: header
      name: X-Project-Id
      description: Project ID scope header
    RunContext:
      type: apiKey
      in: header
      name: X-Run-Id
      description: Optional run ID for correlation
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: not_found
        message:
          type: string
          example: Task not found
    ValidationError:
      type: object
      required: [error, details]
      properties:
        error:
          type: string
          enum: [validation_error]
        details:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              path:
                type: array
                items:
                  type: string
              message:
                type: string
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true

    # --- Project ---
...
    KnowledgeEntity:
      type: object
      properties:
        id:
          type: string
          example: ent_abc123
        projectId:
          type: string
        slug:
          type: string
        name:
          type: string
        entityType:
          $ref: "#/components/schemas/EntityType"
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateEntity:
      type: object
      required: [slug, name]
      properties:
        slug:
          type: string
          minLength: 1
          maxLength: 100
          pattern: ^[a-z0-9-]+$
        name:
          type: string
          minLength: 1
          maxLength: 200
        entityType:
          $ref: "#/components/schemas/EntityType"
        description:
          type: string
          maxLength: 5000
    KnowledgeFact:
      type: object
      properties:
        id:
          type: string
          example: fact_abc123
        entityId:
          type: string
        content:
          type: string
        category:
          $ref: "#/components/schemas/FactCategory"
        sourceAgent:
          type: string
          nullable: true
        sourceTask:
          type: string
          nullable: true
        supersededBy:
          type: string
          nullable: true
        accessCount:
          type: integer
        lastAccessed:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateFact:
      type: object
      required: [content, category]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 10000
        category:
          $ref: "#/components/schemas/FactCategory"
        sourceTask:
          type: string

    # --- Notification ---
...
    EntityType:
      type: string
      enum: [project, area, archive]
      default: area
    FactCategory:
      type: string
      enum: [decision, status, milestone, issue, relationship, convention, observation]
...
paths:
...
  /api/memory/knowledge:
    post:
      tags: [Memory]
      summary: Create knowledge entity
      description: Requires projectId.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntity"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeEntity"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "409":
          description: Duplicate slug in project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [Memory]
      summary: List knowledge entities
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
          description: Auto-scoped from auth if present
        - name: entityType
          in: query
          schema:
            $ref: "#/components/schemas/EntityType"
      responses:
        "200":
          description: Entity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeEntity"
  /api/memory/knowledge/search:
    get:
      tags: [Memory]
      summary: Full-text search across facts
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: projectId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "400":
          description: Missing query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/memory/knowledge/{id}:
    get:
      tags: [Memory]
      summary: Get entity summary
      description: Project-scoped when caller has a projectId.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeEntity"
        "404":
          description: Not found or project mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/memory/knowledge/{id}/facts:
    get:
      tags: [Memory]
      summary: List entity facts
      description: Returns scored facts. Project-scoped.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Fact list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeFact"
        "404":
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: [Memory]
      summary: Write fact
      description: sourceAgent is auto-tagged from the caller's agent ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFact"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeFact"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/memory/knowledge/{entityId}/facts/{factId}/supersede:
    post:
      tags: [Memory]
      summary: Supersede a fact
      description: Replaces an existing fact with an updated version.
      parameters:
        - name: entityId
          in: path
          required: true
          schema:
            type: string
        - name: factId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFact"
      responses:
        "201":
          description: New fact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeFact"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Entity or fact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Fact already superseded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/memory/knowledge/{id}/summarize:
    post:
      tags: [Memory]
      summary: Trigger summary regeneration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Summary regenerated
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Entity or project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/memory/worklog:
    get:
      tags: [Memory]
      summary: Read work log entries
      description: Requires both agentId and projectId.
      parameters:
        - name: agentId
          in: query
          schema:
            type: string
          description: Defaults to caller's agent ID
      responses:
        "200":
          description: Work log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      type: object
        "400":
          description: Missing agentId or projectId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: [Memory]
      summary: Append to work log
      description: Agents can only write their own log. Admins must provide agentId in body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 50000
      responses:
        "201":
          description: Appended
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  filename:
                    type: string
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/memory/lessons:
    get:
      tags: [Memory]
      summary: Read lessons file
      description: Requires both agentId and projectId.
      parameters:
        - name: agentId
          in: query
          schema:
            type: string
          description: Defaults to caller's agent ID
      responses:
        "200":
          description: Lessons content
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
        "400":
          description: Missing agentId or projectId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: [Memory]
      summary: Append to lessons file
      description: Agents can only write their own lessons. Admins must provide agentId in body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 10000
      responses:
        "201":
          description: Appended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ============================================================
  # Cost
  # ============================================================
...